Added CONTRIBUTING.md
The CONTRIBUTING.md file contains all of the information that a user or
developer might need to contribute to the project.  While it is
primarily written for developers who need technical knowledge about how
to contribute code to the project, it also covers the procedures needed
by regular users who simply integration symfony-cron with their
projects.